refactor: separate large files into focused hooks for better maintainability
