+++ components/ProgressDevelopment/DevelopmentProgress.tsx
@@ -1,9 +1,9 @@
 import { UserProgress } from '@/app/(tabs)/home/home.types';
-import { useAuth } from '@/context/AuthContext';
+import { useProgressTracking } from '@/hooks/progressHooks/useProgressTracking';
 import tw from '@/lib/design/tw';
-import { ProgressService } from '@/lib/services/progressService';
-import React, { useCallback, useEffect, useState } from 'react';
-import { Alert, Pressable, Text, View } from 'react-native';
+import React from 'react';
+import { useTranslation } from 'react-i18next';
+import { Pressable, Text, View } from 'react-native';
 
 interface DevelopmentProgressProps {
   onProgressUpdate?: (progress: UserProgress) => void;
@@ -12,65 +12,16 @@ interface DevelopmentProgressProps {
 export const DevelopmentProgress: React.FC<DevelopmentProgressProps> = ({
   onProgressUpdate,
 }) => {
-  const { user } = useAuth();
-  const [progress, setProgress] = useState<UserProgress | null>(null);
-  const [loading, setLoading] = useState(true);
-
-  const loadProgress = useCallback(async () => {
-    if (!user?.uid) return;
-
-    try {
-      setLoading(true);
-      const userProgress = await ProgressService.getProgress(user.uid);
-      if (userProgress) {
-        setProgress(userProgress);
-        onProgressUpdate?.(userProgress);
-      } else {
-        const newProgress = await ProgressService.initializeProgress(user.uid);
-        setProgress(newProgress);
-        onProgressUpdate?.(newProgress);
-      }
-    } catch (error) {
-      console.error('Error loading progress:', error);
-    } finally {
-      setLoading(false);
-    }
-  }, [user?.uid, onProgressUpdate]);
-
-  useEffect(() => {
-    if (user?.uid) {
-      loadProgress();
-    }
-  }, [user?.uid, loadProgress]);
-
-  const updateProgress = async (area: keyof UserProgress, newValue: number) => {
-    if (!user?.uid || !progress) return;
-
-    try {
-      await ProgressService.updateProgress(user.uid, area, newValue);
-
-      const updatedProgress = { ...progress, [area]: newValue };
-      setProgress(updatedProgress);
-      onProgressUpdate?.(updatedProgress);
-
-      console.log(`✅ Progress updated for ${area}: ${newValue}`);
-    } catch (error) {
-      console.error(`❌ Error updating progress for ${area}:`, error);
-      Alert.alert('Error', 'Failed to update progress. Please try again.');
-    }
-  };
-
-  const getProgressColor = (value: number) => {
-    if (value < 30) return 'bg-red-100 text-red-700';
-    if (value < 70) return 'bg-yellow-100 text-yellow-700';
-    return 'bg-green-100 text-green-700';
-  };
-
-  const getProgressLabel = (value: number) => {
-    if (value < 30) return 'Beginner';
-    if (value < 70) return 'Intermediate';
-    return 'Advanced';
-  };
+  const { t } = useTranslation();
+  const {
+    progress,
+    loading,
+    updateProgress,
+    getProgressColor,
+    getProgressLabel,
+  } = useProgressTracking({
+    onProgressUpdate,
+  });
 
   if (loading) {
     return (
@@ -154,7 +105,7 @@ export const DevelopmentProgress: React.FC<DevelopmentProgressProps> = ({
                 updateProgress(key, Math.min(100, progress[key] + 5))
               }
             >
-              <Text style={tw`text-green-700 text-sm font-medium`}>+5</Text>
+              <Text style={tw`text-blue-700 text-sm font-medium`}>+5</Text>
             </Pressable>
           </View>
         </View>
